Allow defining Network Version
During working on a project and testing it on my custom test network,
trying to create a transaction failed due to using a different network
version than the standard Ark network.

This change adds a new optional version argument to createTransaction
which can be used to specify the address version that is used to
validate the passed in address.
The change is fully backwards compatible with prior versions of the
library.